+++ src/Serval.Translation/Contracts/WordGraphArcDto.cs
@@ -4,9 +4,9 @@ public class WordGraphArcDto
 {
     public int PrevState { get; set; }
     public int NextState { get; set; }
-    public float Score { get; set; }
+    public double Score { get; set; }
     public IList<string> TargetTokens { get; set; } = default!;
-    public IList<float> Confidences { get; set; } = default!;
+    public IList<double> Confidences { get; set; } = default!;
     public int SourceSegmentStart { get; set; }
     public int SourceSegmentEnd { get; set; }
     public IList<AlignedWordPairDto> Alignment { get; set; } = default!;
